fix(hearth): missing import, MSVC warning, and test coverage for network sink settings
Real-app verification (--shot against both mockups) caught a genuine
QML runtime bug: NetworkSinkOnlyOnSink.qml used Button without
importing QtQuick.Controls, which the AOT compiler did not flag.
ac3hearth_qmltests' stricter warning set caught a signed/unsigned
comparison in network_controller.cpp.

Adds Catch2 coverage for push_sink_settings()/push_sink_identify()
(capability gating, cache untouched on a failed send) and a Qt Quick
Test suite for the new pages, which needed network_controller.cpp/.hpp
added to ac3hearth_qmltests' own sources - the first tst_*.qml here to
need NetworkController at all.

+++ apps/hearth/ui/network_controller.cpp
@@ -248,7 +248,7 @@ struct LayoutFields {
     routing.reserve(static_cast<qsizetype>(fields.labels.size()));
     const std::optional<ac3::render::Routing> parsed_routing =
         settings.routing ? ac3::render::Routing::parse(*settings.routing, outputs) : std::nullopt;
-    for (std::size_t slot = 0; slot < fields.labels.size(); ++slot) {
+    for (std::size_t slot = 0; slot < static_cast<std::size_t>(fields.labels.size()); ++slot) {
         routing.push_back(parsed_routing ? parsed_routing->output_of(slot) : ac3::render::Routing::kUnassigned);
     }
     map[QStringLiteral("routing")] = routing;
